Parser - when parsing an integer number, a Long will only be created if the number is too big for an Int. This is to match csharp.Parser - when parsing an integer number, a Long will only be created if the number is too big for an Int. This is to match csharp.

+++ Tests/ExpressionBindTests.cs
@@ -13,10 +13,10 @@ namespace ExpressionParserUnitTests
         {
             var host = new TestClass(null, 5);
             var binding = new ExpressionBind(host, "TestInt * 2");
-            Assert.AreEqual(10, (long)binding.Result);
+            Assert.AreEqual(10, (int)binding.Result);
             
             host.TestInt++;
-            Assert.AreEqual(12, (long)binding.Result);
+            Assert.AreEqual(12, (int)binding.Result);
         }
 
         [TestMethod]
@@ -24,10 +24,10 @@ namespace ExpressionParserUnitTests
         {
             var host = new TestClass(null, 5);
             var binding = new ExpressionBind(host, "TestInt * 2");
-            Assert.AreEqual(10, (int)(long)binding.Result);
+            Assert.AreEqual(10, (int)binding.Result);
 
             host.TestInt++;
-            Assert.AreEqual(12, (int)(long)binding.Result);
+            Assert.AreEqual(12, (int)binding.Result);
         }
 
         [TestMethod]
